# static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "sean",
        "version": "1.0.0",
        "title": "Seans-Python3-Flask-rest-Boilerplate",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [{
        "url": "/"
    }],
...
    "paths": {
...
        "/locataire/appart/all": {
            "get": {
                "tags": [
                    "Locataire request"
                ],
                "summary": "Permet de voir la liste des appartements",
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "Cette erreur se produit quand il y'a un problème côté client"
                    },
                    "401": {
                        "description": "Cette erreur se produit quand l'utilisateur n'est pas autorisé"
                    },
                    "404": {
                        "description": "Cette erreur se produit quand la ressource demandée n'a pas été trouvée"
                    }
                }
            }
        },
        "/locataire/appart/search": {
            "post": {
                "tags": [
                    "Locataire request"
                ],
                "summary": "research an appart",
                "requestBody": {
                    "description": "Search an appart that math some criteria",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/search_appart"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/search_appart"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/locataire/appart/group/create": {
            "post": {
                "tags": [
                    "Locataire request"
                ],
                "security": [{
                    "bearerAuth": []
                }],
                "summary": "research an appart",
                "requestBody": {
                    "description": "Search an appart that math some criteria",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/group"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/group"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/locataire/appart/group/join": {
            "post": {
                "tags": [
                    "Locataire request"
                ],
                "security": [{
                    "bearerAuth": []
                }],
                "summary": "research an appart",
                "requestBody": {
                    "description": "Search an appart that math some criteria",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/join"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/join"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/locataire/appart/group/leave": {
            "post": {
                "tags": [
                    "Locataire request"
                ],
                "security": [{
                    "bearerAuth": []
                }],
                "summary": "leave group",
                "requestBody": {
                    "description": "leave group",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/join"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/join"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/locataire/appart/groups/my": {
            "get": {
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Locataire request"
                ],
                "summary": "Permet de voir la liste des groupes auxquel j'appartient",
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "Cette erreur se produit quand il y'a un problème côté client"
                    },
                    "401": {
                        "description": "Cette erreur se produit quand l'utilisateur n'est pas autorisé"
                    },
                    "404": {
                        "description": "Cette erreur se produit quand la ressource demandée n'a pas été trouvée"
                    }
                }
            }
        },
        "/locataire/groups/for/appart": {
            "get": {
                "parameters": [{
                    "name": "appart_id",
                    "in": "query",
                    "required": false,
                    "description": "l'id de l'appartement",
                    "type": "string"
                }],
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Locataire request"
                ],
                "summary": "Permet de voir la liste des groupes par appartement",
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "Cette erreur se produit quand il y'a un problème côté client"
                    },
                    "401": {
                        "description": "Cette erreur se produit quand l'utilisateur n'est pas autorisé"
                    },
                    "404": {
                        "description": "Cette erreur se produit quand la ressource demandée n'a pas été trouvée"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "group": {
                "type": "object",
                "properties": {
                    "appart_id": {
                        "type": "string",
                        "format": "string"
                    },
                    "groupe_name": {
                        "type": "string",
                        "format": "string"
                    }
                }
            },
...
            "join": {
                "type": "object",
                "properties": {
                    "group_id": {
                        "type": "string",
                        "format": "string"
                    }
                }
            },
            "search_appart": {
                "type": "object",
                "properties": {
                    "country": {
                        "type": "string",
                        "format": "string"
                    },
                    "town": {
                        "type": "string",
                        "format": "string"
                    },
                    "district": {
                        "type": "string",
                        "format": "string"
                    },
                    "specifications": {
                        "type": "object",
                        "format": "object",
                        "example": {
                            "room": 1,
                            "shower": 1,
                            "toilet": 1,
                            "living room": 1,
                            "kitchen": 1,
                            "veranda": 1,
                            "backyard": 1,
                            "garage": 1
                        }
                    },
                    "min_price": {
                        "type": "int",
                        "format": "string",
                        "example": 25000
                    },
                    "max_price": {
                        "type": "int",
                        "format": "string",
                        "example": 30000
                    }
                }
            },
...
        }
    }
}